Blackjack counts the score of a player's hand of cards.
This originally had a rich and detailed git history of commits associated with it, but then something weird happened and it pulled it a thousand commits from another repository and made the repo hold like a thousand files and -- it was unnecessary and annoying.  So I copied the files and restarted the git repo.  I don't think anyone will miss that history, to be honest.  It works now.  That's what counts.  And it's small enough that making changes won't get ugly.